feat: Add support for adding and displaying reviewers in gerrit
This commit adds two new functions to the git_gerrit.sh script. The first function, gt-gerrit-reviewers-add, allows users to add a reviewer's email address to a list stored in a file. The second function, gt-gerrit-reviewers, displays the list of reviewers from the file. These functions provide a convenient way to manage reviewers for gerrit patches.